feat(layout): unify header and footer
